Update function3_draft.py
Here is a new version of the "function 3" code that aims at solving seating.

For the moment, this function extracts the zeros and ones matrix from a txt file called txt_airline_seating.txt. 
Still to do : Once integrated in the main code it will use the variable binary_db

In this version, the function solves seating properly until Exceptions occurs.
Still to do : I have to find a way to allow partysizes to be separated and to accept partysizes greater than 4.

The next_new_booking list works properly but it gives seat numbers (from 1 to 60) instead of rows and col numbers. 
Still to do :I think I will simply write additional lines of codes that convert for example "5" into "2,A"